refactor(ocr): 抽取验证码自动填写公共状态机，合并 id/zhw 两份拷贝
- 新增 captcha-autofill.ts：识别/填写/刷新重试/事件监听的状态机只保留一份，
  各登录页通过 locate/onClickCapture/recognize/warmup 注入差异点
- id-captcha.ts 与 zhjw-captcha-ocr.ts 只保留各自的页面定位逻辑
- 移除无人引用的 warmupZhjwOcr / ocrZhjw 导出（已验证无外部消费方）
- 已知内部差异：zhjw 此前在输入框缺失时也会给图片绑 load 监听，
  统一后改为图片+输入框都在场才绑定，MutationObserver 重入可覆盖该间隙
